api/requirements.txt
Flask
flask-cors
moviepy
whisper
openai
python-dotenv
